+++ src/core/component-manager.js
@@ -1,0 +1,58 @@
+/**
+ * This function adds one to its input.
+ * @param {number} input any number
+ * @returns {number} that number, plus one.
+ */
+class ComponentManager {
+    constructor() {
+        this.init()
+    }
+    
+    init() {
+        this.components = new Map()
+    }
+    
+    newComponent(componentId) {
+        let component = this.components.get(componentId)
+        
+        if (component === null || component === undefined) {
+            return null
+        }
+        
+        switch (typeof component) {
+            case 'function':
+                return new component()
+            case 'object'  : {
+                return ((component) => {
+                    let ret = {}
+                    
+                    Object.keys(component).forEach(key => ret[key] = component[key])
+                    
+                    return ret
+                })(component)
+            }
+            default:
+                return component
+        }
+    }
+    
+    registerComponent(component) {
+        if (component === null || component === undefined) {
+            throw TypeError('component cannot be null or undefined.')
+        }
+        
+        const max = Math.max(...this.components.keys())
+        
+        const id = max === undefined || max === null || max === -Infinity ? 1 : max === 0 ? 1 : max * 2
+
+        this.components.set(id, component)
+
+        return id
+    }
+    
+    getComponents() {
+        return this.components
+    }
+}
+
+export { ComponentManager }