+++ src/index.js
@@ -1,0 +1,7 @@
+import { EntityManager }             from './core/entity-manager'
+import { EntityFactory }             from './core/entity-factory'
+import { SystemManager, SystemType } from './core/system-manager'
+import { ComponentManager }          from './core/component-manager'
+import { EventHandler }              from './core/event-handler'
+
+export { EntityManager, EntityFactory, SystemManager, SystemType, ComponentManager, EventHandler }