+++ test/entity/manager/selector-type.js
@@ -1,8 +1,0 @@
-import { expect }       from 'chai';
-import { SelectorType } from '../../../src/core/entity';
-
-describe('SelectorType', function() {
-    it('is an object', () => {
-        expect(SelectorType).to.be.an('object');
-    });
-});